document new configuration
